+++ src/Http/Controllers/Api/Controller.php
@@ -8,6 +8,9 @@ use App\Http\Controllers\Controller as AppController;
  * Team controller.
  *
  * @since 0.1.0
+ *
+ * @property class-string $model
+ * @property string $ability
  */
 class Controller extends AppController
 {