+++ src/Http/Middleware/HasTeamPermission.php
@@ -1,0 +1,22 @@
+<?php
+
+namespace BrilliantPortal\Framework\Http\Middleware;
+
+use Closure;
+use Illuminate\Http\Request;
+
+class HasTeamPermission
+{
+    public function handle(Request $request, Closure $next, string $permission)
+    {
+        if (! $request->user()) {
+            abort(401, 'Unauthenticated');
+        }
+
+        if (! $request->user()->hasTeamPermission($request->user()->currentTeam, $permission)) {
+            abort(403, 'Forbidden');
+        }
+
+        return $next($request);
+    }
+}