+++ stubs/tests/Api/DocumentationTest.stub.php
@@ -6,6 +6,8 @@ use App\Models\Team;
 use App\Models\User;
 use BrilliantPortal\Framework\Framework;
 use Illuminate\Foundation\Testing\RefreshDatabase;
+use Inertia\Inertia;
+use Inertia\Testing\AssertableInertia as Assert;
 use Laravel\Jetstream\Features;
 use Tests\TestCase;
 
@@ -16,11 +18,11 @@ class DocumentationTest extends TestCase
     public function test_logged_out_users_cant_see_api_docs()
     {
         if (! Features::hasTeamFeatures()) {
-            return $this->markTestSkipped('Teams support is not enabled.');
+            $this->marktestSkipped('Teams support is not enabled.');
         }
 
         if (! Features::hasApiFeatures()) {
-            return $this->markTestSkipped('API support is not enabled.');
+            $this->marktestSkipped('API support is not enabled.');
         }
 
         $this
@@ -31,11 +33,11 @@ class DocumentationTest extends TestCase
     public function test_non_admin_users_cant_see_api_docs()
     {
         if (! Features::hasTeamFeatures()) {
-            return $this->markTestSkipped('Teams support is not enabled.');
+            $this->marktestSkipped('Teams support is not enabled.');
         }
 
         if (! Features::hasApiFeatures()) {
-            return $this->markTestSkipped('API support is not enabled.');
+            $this->marktestSkipped('API support is not enabled.');
         }
 
         $team = Team::factory()->create();
@@ -54,21 +56,30 @@ class DocumentationTest extends TestCase
     public function test_admin_users_can_see_api_docs()
     {
         if (! Features::hasTeamFeatures()) {
-            return $this->markTestSkipped('Teams support is not enabled.');
+            $this->marktestSkipped('Teams support is not enabled.');
         }
 
         if (! Features::hasApiFeatures()) {
-            return $this->markTestSkipped('API support is not enabled.');
+            $this->marktestSkipped('API support is not enabled.');
         }
 
         Framework::addApiAuthMechanism();
 
         $user = User::factory()->withPersonalTeam()->create();
 
-        $this
+        $response = $this
             ->actingAs($user)
             ->get('/dashboard/api-documentation')
-            ->assertOk()
-            ->assertViewIs('brilliant-portal-framework::api.documentation');
+            ->assertOk();
+
+
+        if (class_exists(Inertia::class)) {
+            $response->assertInertia(fn (Assert $page) => $page
+                ->component('API/Documentation')
+                ->has('spec')
+            );
+        } else {
+            $response->assertViewIs('brilliant-portal-framework::api.documentation');
+        }
     }
 }