+++ stubs/tests/EnsureHasNoTeamTest.stub.php
@@ -5,6 +5,8 @@ namespace Tests\Feature;
 use App\Models\User;
 use App\Providers\RouteServiceProvider;
 use Illuminate\Foundation\Testing\RefreshDatabase;
+use Inertia\Inertia;
+use Inertia\Testing\AssertableInertia as Assert;
 use Laravel\Jetstream\Features;
 use Tests\TestCase;
 
@@ -15,7 +17,7 @@ class EnsureHasNoTeamTest extends TestCase
     public function test_user_without_teams_can_create_first_team()
     {
         if (! Features::hasTeamFeatures()) {
-            return $this->markTestSkipped('Teams support is not enabled.');
+            $this->marktestSkipped('Teams support is not enabled.');
         }
 
         $user = User::factory()->create();
@@ -24,16 +26,22 @@ class EnsureHasNoTeamTest extends TestCase
             'user_id' => $user->id,
         ]);
 
-        $this
+        $response = $this
             ->actingAs($user)
             ->get(route('brilliant-portal-framework.teams.create-first'))
-            ->assertViewIs('brilliant-portal-framework::teams.create-first');
+            ->assertOk();
+
+        if (class_exists(Inertia::class)) {
+            $response->assertInertia(fn (Assert $page) => $page->component('Teams/CreateFirst'));
+        } else {
+            $response->assertViewIs('brilliant-portal-framework::teams.create-first');
+        }
     }
 
     public function test_user_without_teams_can_see_already_invited()
     {
         if (! Features::hasTeamFeatures()) {
-            return $this->markTestSkipped('Teams support is not enabled.');
+            $this->marktestSkipped('Teams support is not enabled.');
         }
 
         $user = User::factory()->create();
@@ -42,16 +50,22 @@ class EnsureHasNoTeamTest extends TestCase
             'user_id' => $user->id,
         ]);
 
-        $this
+        $response = $this
             ->actingAs($user)
             ->get(route('brilliant-portal-framework.teams.already-invited'))
-            ->assertViewIs('brilliant-portal-framework::teams.already-invited');
+            ->assertOk();
+
+        if (class_exists(Inertia::class)) {
+            $response->assertInertia(fn (Assert $page) => $page->component('Teams/AlreadyInvited'));
+        } else {
+            $response->assertViewIs('brilliant-portal-framework::teams.already-invited');
+        }
     }
 
     public function test_user_with_teams_cant_create_first_team()
     {
         if (! Features::hasTeamFeatures()) {
-            return $this->markTestSkipped('Teams support is not enabled.');
+            $this->marktestSkipped('Teams support is not enabled.');
         }
 
         $user = User::factory()->withPersonalTeam()->create();
@@ -70,7 +84,7 @@ class EnsureHasNoTeamTest extends TestCase
     public function test_user_with_teams_cant_see_already_invited()
     {
         if (! Features::hasTeamFeatures()) {
-            return $this->markTestSkipped('Teams support is not enabled.');
+            $this->marktestSkipped('Teams support is not enabled.');
         }
 
         $user = User::factory()->withPersonalTeam()->create();