+++ stubs/vite-inertia.config.js
@@ -1,5 +1,6 @@
 import { defineConfig } from 'vite';
 import laravel from 'laravel-vite-plugin';
+import vue from '@vitejs/plugin-vue';
 
 let host = 'localhost.test';
 
@@ -11,6 +12,15 @@ export default defineConfig({
             ],
             refresh: true,
             valetTls: host,
+        }),
+
+        vue({
+            template: {
+                transformAssetUrls: {
+                    base: null,
+                    includeAbsolute: false,
+                },
+            },
         }),
     ],
 });