+++ stubs/vite-livewire.config.js
@@ -1,8 +1,5 @@
-import fs from 'fs';
 import { defineConfig } from 'vite';
 import laravel from 'laravel-vite-plugin';
-import {homedir} from 'os';
-import {resolve} from 'path';
 import livewire from '@defstudio/vite-livewire-plugin';
 
 let host = 'localhost.test';
@@ -15,34 +12,11 @@ export default defineConfig({
                 'resources/js/app.js',
             ],
             refresh: true,
+            valetTls: host,
         }),
 
         livewire({
             refresh: ['resources/css/app.css'],
         }),
     ],
-
-    server: detectServerConfig(host),
 });
-
-function detectServerConfig(host) {
-    let keyPath = resolve(homedir(), `.config/valet/Certificates/${host}.key`)
-    let certificatePath = resolve(homedir(), `.config/valet/Certificates/${host}.crt`)
-
-    if (!fs.existsSync(keyPath)) {
-        return {}
-    }
-
-    if (!fs.existsSync(certificatePath)) {
-        return {}
-    }
-
-    return {
-        hmr: {host},
-        host,
-        https: {
-            key: fs.readFileSync(keyPath),
-            cert: fs.readFileSync(certificatePath),
-        },
-    }
-}